Added extern scope for segfault handler registration.
