+++ compiler-rt/lib/mpk_untrusted/CMakeLists.txt
@@ -5,6 +5,7 @@ include_directories(..)
 set(MPK_UNTRUSTED_SOURCES
     alloc_site_handler.cpp
     mpk_fault_handler.cpp
+    mpk_untrusted.cpp
     )
 
 set(MPK_UNTRUSTED_HEADERS
